Move chat router services to own file
